// back-end/src/shared/enum.ts
export enum PassionEnum {
    MEDIUM = 'Medium',
    HIGH = 'High',
    LOW = 'Low',
    VERY_HIGH = 'Very_High',
}

export enum CollectionEmun {
    USER = 'User',
    HOBBIES = 'Hobbies'
}

export enum StatusHttpCodeEnum {
    OK = 200,
    CREATED = 201,
    BAD_REQUEST = 400,
    FORBIDDEN = 403,
    NOT_FOUND = 404,
    SERVER_ERROR = 500,
}